# api-contract.yml
openapi: 3.0.0
info: 
  title: United Nest API
  description: This is the API Documentation of the United Nest REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/
    description: DEV Server
  - url: https://www.united-nest.com/
    description: PROD Server
paths:  # TODO: add authentication methods for required endpoints
...
  /api/v1/connections/following:
    get:
      description: 'Return all exisiting users whom the current user is following'
      tags:
        - Connections
      responses:
        '200':
          description: 'Ok, success'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/statusSchema'
                  results:
                    type: integer
                  data:
                    type: object
                    properties:
                      connections:
                        type: array
                        items:
                          $ref: '#/components/schemas/connectionDocument'
                required:
                  - status
                  - data
        '400':
          description: 'Client Side Error - Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
 
  # Get All Followers endpoint
  /api/v1/connections/followers:
    get:
      description: 'Return all exisiting users who are followers of current user'
      tags:
        - Connections
      responses:
        '200':
          description: 'Ok, success'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/statusSchema'
                  results:
                    type: integer
                  data:
                    type: object
                    properties:
                      connections:
                        type: array
                        items:
                          $ref: '#/components/schemas/connectionDocument'
                required:
                  - status
                  - data
        '400':
          description: 'Client Side Error - Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
 
  # Unfollower User endpoint
...
components:
...
  schemas: # TODO: add examples to all schemas 😢
...
    error:  # TODO: Change the Schema once you differentiate DEV & PROD errors in errorController.js
      type: object
      properties:
        status: 
          $ref: '#/components/schemas/statusSchema'
        message:
          type: string

    # ** Login payload with username, password **
...
    statusSchema:
      type: string
      enum:
        - success
        - fail
        - error

    # ** token (JWT) **
...
    connectionDocument:
      type: object
      properties:
        requestSender:
          type: string
        requestReceiver:
          type: string
        requestSentTime:
          type: string
        status:
          type: string
          enum: ['success', 'fail', 'error']
      required:
        - requestSender
        - requestReceiver
        - requestSentTime
        - status

    # ** Schema for Delete Request Response (204) **
...
  securitySchemes:

    # Authorization header: Bearer [TOKEN]
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization header in the format 'Bearer [TOKEN]'"

    # Cookie Auth (unsafe, prone to CSRF attacks)
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: "JWT Token in the cookie"